% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_thumbnail.R
\name{get.rng.from.line.number}
\alias{get.rng.from.line.number}
\title{An RStudio addin for inserting newest image into a blogdown post}
\usage{
get.rng.from.line.number(line.number)
}
\description{
insert.thumbnail()
}
