% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_newest.R
\name{insert.newest.image.from.desktop}
\alias{insert.newest.image.from.desktop}
\title{An RStudio addin for inserting newest image into a blogdown post}
\usage{
insert.newest.image.from.desktop()
}
\description{
insert.newest.image.from.desktop()
}
